requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
dj-database-url==2.3.0
Django==5.2.1
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
